remove nlp and based on sentence-transformers
