--- price_recommender/nlp/distilbert-base-nli-stsb-mean-tokens/0_Transformer/config.json
@@ -1,21 +1,0 @@
-{
-  "activation": "gelu",
-  "architectures": [
-    "DistilBertModel"
-  ],
-  "attention_dropout": 0.1,
-  "dim": 768,
-  "dropout": 0.1,
-  "hidden_dim": 3072,
-  "initializer_range": 0.02,
-  "max_position_embeddings": 512,
-  "model_type": "distilbert",
-  "n_heads": 12,
-  "n_layers": 6,
-  "pad_token_id": 0,
-  "qa_dropout": 0.1,
-  "seq_classif_dropout": 0.2,
-  "sinusoidal_pos_embds": false,
-  "tie_weights_": true,
-  "vocab_size": 30522
-}
