--- price_recommender/nlp/distilbert-base-nli-stsb-mean-tokens/0_Transformer/pytorch_model.bin
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:597080ab75b867ef3ba15633bdc5c4cd55c3563ee92840e7af826464b970a0f1
-size 265473819
