--- price_recommender/nlp/distilbert-base-nli-stsb-mean-tokens/0_Transformer/sentence_bert_config.json
@@ -1,3 +1,0 @@
-{
-  "max_seq_length": 128
-}
