--- price_recommender/nlp/distilbert-base-nli-stsb-mean-tokens/1_Pooling/config.json
@@ -1,7 +1,0 @@
-{
-  "word_embedding_dimension": 768,
-  "pooling_mode_cls_token": false,
-  "pooling_mode_mean_tokens": true,
-  "pooling_mode_max_tokens": false,
-  "pooling_mode_mean_sqrt_len_tokens": false
-}
