--- price_recommender/nlp/distilbert-base-nli-stsb-mean-tokens/modules.json
@@ -1,14 +1,0 @@
-[
-  {
-    "idx": 0,
-    "name": "0",
-    "path": "0_Transformer",
-    "type": "nlp.models.Transformer"
-  },
-  {
-    "idx": 1,
-    "name": "1",
-    "path": "1_Pooling",
-    "type": "nlp.models.Pooling"
-  }
-]
