--- price_recommender/nlp/models/__init__.py
@@ -1,3 +1,0 @@
-from .DistilBERT import *
-from .Pooling import *
-from .Transformer import *
